Refactor connection get and set methods to enforce propper logging

+++ meteor/src/render_system.cpp
@@ -160,9 +160,9 @@ namespace meteor::render{
 			
 			int y_offset = 80 + 175 * (game_instance.m_user_index - 1);	// Mimic server alignment for convenience
 
-			render_rtt_graph(HUD_OFFSET + Vector2(300, y_offset), conn, -500.0f);			// seconds to ms, 1px per 2ms
-			render_send_bytes_hist(HUD_OFFSET + Vector2(420, y_offset), conn, -0.25f);		// 1px per 4 bytes
-			render_recieve_bytes_hist(HUD_OFFSET + Vector2(540, y_offset), conn, -0.25f);	// 1px per 4 bytes
+			render_rtt_graph(HUD_OFFSET + Vector2(300, (float)y_offset), conn, -500.0f);			// seconds to ms, 1px per 2ms
+			render_send_bytes_hist(HUD_OFFSET + Vector2(420, (float)y_offset), conn, -0.25f);		// 1px per 4 bytes
+			render_recieve_bytes_hist(HUD_OFFSET + Vector2(540, (float)y_offset), conn, -0.25f);	// 1px per 4 bytes
 
 			render_connection_stats(Vector2i(80, 20), conn);
 		}
